+++ aws-java-sdk-mediapackagevod/src/main/java/com/amazonaws/services/mediapackagevod/model/DashManifest.java
@@ -26,6 +26,12 @@ import com.amazonaws.protocol.ProtocolMarshaller;
 @Generated("com.amazonaws:aws-java-sdk-code-generator")
 public class DashManifest implements Serializable, Cloneable, StructuredPojo {
 
+    /**
+     * Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like
+     * SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate
+     * elements are combined and presented at the AdaptationSet level.
+     */
+    private String manifestLayout;
     /** An optional string to include in the name of the manifest. */
     private String manifestName;
     /** Minimum duration (in seconds) that a player will buffer media before starting the presentation. */
@@ -37,6 +43,73 @@ public class DashManifest implements Serializable, Cloneable, StructuredPojo {
     private String profile;
 
     private StreamSelection streamSelection;
+
+    /**
+     * Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like
+     * SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate
+     * elements are combined and presented at the AdaptationSet level.
+     * 
+     * @param manifestLayout
+     *        Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL,
+     *        elements like SegmentTemplate and ContentProtection are included in each Representation. When set to
+     *        COMPACT, duplicate elements are combined and presented at the AdaptationSet level.
+     * @see ManifestLayout
+     */
+
+    public void setManifestLayout(String manifestLayout) {
+        this.manifestLayout = manifestLayout;
+    }
+
+    /**
+     * Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like
+     * SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate
+     * elements are combined and presented at the AdaptationSet level.
+     * 
+     * @return Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL,
+     *         elements like SegmentTemplate and ContentProtection are included in each Representation. When set to
+     *         COMPACT, duplicate elements are combined and presented at the AdaptationSet level.
+     * @see ManifestLayout
+     */
+
+    public String getManifestLayout() {
+        return this.manifestLayout;
+    }
+
+    /**
+     * Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like
+     * SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate
+     * elements are combined and presented at the AdaptationSet level.
+     * 
+     * @param manifestLayout
+     *        Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL,
+     *        elements like SegmentTemplate and ContentProtection are included in each Representation. When set to
+     *        COMPACT, duplicate elements are combined and presented at the AdaptationSet level.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see ManifestLayout
+     */
+
+    public DashManifest withManifestLayout(String manifestLayout) {
+        setManifestLayout(manifestLayout);
+        return this;
+    }
+
+    /**
+     * Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like
+     * SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate
+     * elements are combined and presented at the AdaptationSet level.
+     * 
+     * @param manifestLayout
+     *        Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL,
+     *        elements like SegmentTemplate and ContentProtection are included in each Representation. When set to
+     *        COMPACT, duplicate elements are combined and presented at the AdaptationSet level.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see ManifestLayout
+     */
+
+    public DashManifest withManifestLayout(ManifestLayout manifestLayout) {
+        this.manifestLayout = manifestLayout.toString();
+        return this;
+    }
 
     /**
      * An optional string to include in the name of the manifest.
@@ -203,6 +276,8 @@ public class DashManifest implements Serializable, Cloneable, StructuredPojo {
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
+        if (getManifestLayout() != null)
+            sb.append("ManifestLayout: ").append(getManifestLayout()).append(",");
         if (getManifestName() != null)
             sb.append("ManifestName: ").append(getManifestName()).append(",");
         if (getMinBufferTimeSeconds() != null)
@@ -225,6 +300,10 @@ public class DashManifest implements Serializable, Cloneable, StructuredPojo {
         if (obj instanceof DashManifest == false)
             return false;
         DashManifest other = (DashManifest) obj;
+        if (other.getManifestLayout() == null ^ this.getManifestLayout() == null)
+            return false;
+        if (other.getManifestLayout() != null && other.getManifestLayout().equals(this.getManifestLayout()) == false)
+            return false;
         if (other.getManifestName() == null ^ this.getManifestName() == null)
             return false;
         if (other.getManifestName() != null && other.getManifestName().equals(this.getManifestName()) == false)
@@ -249,6 +328,7 @@ public class DashManifest implements Serializable, Cloneable, StructuredPojo {
         final int prime = 31;
         int hashCode = 1;
 
+        hashCode = prime * hashCode + ((getManifestLayout() == null) ? 0 : getManifestLayout().hashCode());
         hashCode = prime * hashCode + ((getManifestName() == null) ? 0 : getManifestName().hashCode());
         hashCode = prime * hashCode + ((getMinBufferTimeSeconds() == null) ? 0 : getMinBufferTimeSeconds().hashCode());
         hashCode = prime * hashCode + ((getProfile() == null) ? 0 : getProfile().hashCode());