+++ aws-java-sdk-mediapackagevod/src/main/java/com/amazonaws/services/mediapackagevod/model/DashPackage.java
@@ -31,10 +31,24 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
 
     private DashEncryption encryption;
     /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     */
+    private java.util.List<String> periodTriggers;
+    /**
      * Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source
      * segment duration.
      */
     private Integer segmentDurationSeconds;
+    /**
+     * Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     * NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to
+     * TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to
+     * NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.
+     */
+    private String segmentTemplateFormat;
 
     /**
      * A list of DASH manifest configurations.
@@ -125,6 +139,125 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
     }
 
     /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     * 
+     * @return A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media
+     *         Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
+     *         be partitioned into more than one period. If the list contains "ADS", new periods will be created where
+     *         the Asset contains SCTE-35 ad markers.
+     * @see PeriodTriggersElement
+     */
+
+    public java.util.List<String> getPeriodTriggers() {
+        return periodTriggers;
+    }
+
+    /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     * 
+     * @param periodTriggers
+     *        A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media
+     *        Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
+     *        be partitioned into more than one period. If the list contains "ADS", new periods will be created where
+     *        the Asset contains SCTE-35 ad markers.
+     * @see PeriodTriggersElement
+     */
+
+    public void setPeriodTriggers(java.util.Collection<String> periodTriggers) {
+        if (periodTriggers == null) {
+            this.periodTriggers = null;
+            return;
+        }
+
+        this.periodTriggers = new java.util.ArrayList<String>(periodTriggers);
+    }
+
+    /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     * 
+     * <p>
+     * <b>NOTE:</b> This method appends the values to the existing list (if any). Use
+     * {@link #setPeriodTriggers(java.util.Collection)} or {@link #withPeriodTriggers(java.util.Collection)} if you want
+     * to override the existing values.
+     * </p>
+     * 
+     * @param periodTriggers
+     *        A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media
+     *        Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
+     *        be partitioned into more than one period. If the list contains "ADS", new periods will be created where
+     *        the Asset contains SCTE-35 ad markers.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see PeriodTriggersElement
+     */
+
+    public DashPackage withPeriodTriggers(String... periodTriggers) {
+        if (this.periodTriggers == null) {
+            setPeriodTriggers(new java.util.ArrayList<String>(periodTriggers.length));
+        }
+        for (String ele : periodTriggers) {
+            this.periodTriggers.add(ele);
+        }
+        return this;
+    }
+
+    /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     * 
+     * @param periodTriggers
+     *        A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media
+     *        Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
+     *        be partitioned into more than one period. If the list contains "ADS", new periods will be created where
+     *        the Asset contains SCTE-35 ad markers.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see PeriodTriggersElement
+     */
+
+    public DashPackage withPeriodTriggers(java.util.Collection<String> periodTriggers) {
+        setPeriodTriggers(periodTriggers);
+        return this;
+    }
+
+    /**
+     * A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation
+     * Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into
+     * more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad
+     * markers.
+     * 
+     * @param periodTriggers
+     *        A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media
+     *        Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not
+     *        be partitioned into more than one period. If the list contains "ADS", new periods will be created where
+     *        the Asset contains SCTE-35 ad markers.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see PeriodTriggersElement
+     */
+
+    public DashPackage withPeriodTriggers(PeriodTriggersElement... periodTriggers) {
+        java.util.ArrayList<String> periodTriggersCopy = new java.util.ArrayList<String>(periodTriggers.length);
+        for (PeriodTriggersElement value : periodTriggers) {
+            periodTriggersCopy.add(value.toString());
+        }
+        if (getPeriodTriggers() == null) {
+            setPeriodTriggers(periodTriggersCopy);
+        } else {
+            getPeriodTriggers().addAll(periodTriggersCopy);
+        }
+        return this;
+    }
+
+    /**
      * Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source
      * segment duration.
      * 
@@ -165,6 +298,85 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
     }
 
     /**
+     * Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     * NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to
+     * TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to
+     * NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.
+     * 
+     * @param segmentTemplateFormat
+     *        Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     *        NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When
+     *        set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs.
+     *        When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media
+     *        URLs.
+     * @see SegmentTemplateFormat
+     */
+
+    public void setSegmentTemplateFormat(String segmentTemplateFormat) {
+        this.segmentTemplateFormat = segmentTemplateFormat;
+    }
+
+    /**
+     * Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     * NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to
+     * TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to
+     * NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.
+     * 
+     * @return Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     *         NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs.
+     *         When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media
+     *         URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with
+     *         $Number$ media URLs.
+     * @see SegmentTemplateFormat
+     */
+
+    public String getSegmentTemplateFormat() {
+        return this.segmentTemplateFormat;
+    }
+
+    /**
+     * Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     * NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to
+     * TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to
+     * NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.
+     * 
+     * @param segmentTemplateFormat
+     *        Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     *        NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When
+     *        set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs.
+     *        When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media
+     *        URLs.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see SegmentTemplateFormat
+     */
+
+    public DashPackage withSegmentTemplateFormat(String segmentTemplateFormat) {
+        setSegmentTemplateFormat(segmentTemplateFormat);
+        return this;
+    }
+
+    /**
+     * Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     * NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to
+     * TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to
+     * NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.
+     * 
+     * @param segmentTemplateFormat
+     *        Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to
+     *        NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When
+     *        set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs.
+     *        When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media
+     *        URLs.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     * @see SegmentTemplateFormat
+     */
+
+    public DashPackage withSegmentTemplateFormat(SegmentTemplateFormat segmentTemplateFormat) {
+        this.segmentTemplateFormat = segmentTemplateFormat.toString();
+        return this;
+    }
+
+    /**
      * Returns a string representation of this object. This is useful for testing and debugging. Sensitive data will be
      * redacted from this string using a placeholder value.
      *
@@ -180,8 +392,12 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
             sb.append("DashManifests: ").append(getDashManifests()).append(",");
         if (getEncryption() != null)
             sb.append("Encryption: ").append(getEncryption()).append(",");
+        if (getPeriodTriggers() != null)
+            sb.append("PeriodTriggers: ").append(getPeriodTriggers()).append(",");
         if (getSegmentDurationSeconds() != null)
-            sb.append("SegmentDurationSeconds: ").append(getSegmentDurationSeconds());
+            sb.append("SegmentDurationSeconds: ").append(getSegmentDurationSeconds()).append(",");
+        if (getSegmentTemplateFormat() != null)
+            sb.append("SegmentTemplateFormat: ").append(getSegmentTemplateFormat());
         sb.append("}");
         return sb.toString();
     }
@@ -204,9 +420,17 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
             return false;
         if (other.getEncryption() != null && other.getEncryption().equals(this.getEncryption()) == false)
             return false;
+        if (other.getPeriodTriggers() == null ^ this.getPeriodTriggers() == null)
+            return false;
+        if (other.getPeriodTriggers() != null && other.getPeriodTriggers().equals(this.getPeriodTriggers()) == false)
+            return false;
         if (other.getSegmentDurationSeconds() == null ^ this.getSegmentDurationSeconds() == null)
             return false;
         if (other.getSegmentDurationSeconds() != null && other.getSegmentDurationSeconds().equals(this.getSegmentDurationSeconds()) == false)
+            return false;
+        if (other.getSegmentTemplateFormat() == null ^ this.getSegmentTemplateFormat() == null)
+            return false;
+        if (other.getSegmentTemplateFormat() != null && other.getSegmentTemplateFormat().equals(this.getSegmentTemplateFormat()) == false)
             return false;
         return true;
     }
@@ -218,7 +442,9 @@ public class DashPackage implements Serializable, Cloneable, StructuredPojo {
 
         hashCode = prime * hashCode + ((getDashManifests() == null) ? 0 : getDashManifests().hashCode());
         hashCode = prime * hashCode + ((getEncryption() == null) ? 0 : getEncryption().hashCode());
+        hashCode = prime * hashCode + ((getPeriodTriggers() == null) ? 0 : getPeriodTriggers().hashCode());
         hashCode = prime * hashCode + ((getSegmentDurationSeconds() == null) ? 0 : getSegmentDurationSeconds().hashCode());
+        hashCode = prime * hashCode + ((getSegmentTemplateFormat() == null) ? 0 : getSegmentTemplateFormat().hashCode());
         return hashCode;
     }